Finish to add new functionality (search by username or tag)

+++ InstagramKit-Example/InstagramKit-Example/Classes/View/SearchView Cells/IKSearchTagCell.m
@@ -37,7 +37,7 @@
 
 - (void)setMediaCount:(NSInteger)mediaCount
 {
-    [self.labelMediaCount setText:[NSString stringWithFormat:@"%ld", mediaCount]];
+    [self.labelMediaCount setText:[NSString stringWithFormat:@"%ld", (long)mediaCount]];
 }
